combine restriction key handler with edit key handler
