+++ packages/iris-grid/src/key-handlers/index.ts
@@ -2,4 +2,3 @@ export { default as CopyCellKeyHandler } from './CopyCellKeyHandler';
 export { default as CopyKeyHandler } from './CopyKeyHandler';
 export { default as ReverseKeyHandler } from './ReverseKeyHandler';
 export { default as ClearFilterKeyHandler } from './ClearFilterKeyHandler';
-export { default as RestrictedEditKeyHandler } from './RestrictedEditKeyHandler';
